% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobra_functions.R
\name{adj_ortho}
\alias{adj_ortho}
\title{adj_ortho: internal helper for orthogonal projection}
\usage{
adj_ortho(dat, design, loc, reduce)
}
\arguments{
\item{dat}{Transposed or original data (depending on 'reduce')}

\item{design}{Model matrix (batch and other covariates)}

\item{loc}{Indices of batch columns in design}

\item{reduce}{Logical; if TRUE, use different fitting approach}
}
\value{
A batch-effect-adjusted matrix
}
\description{
adj_ortho: internal helper for orthogonal projection
}
\keyword{internal}
