% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BALUT.R
\name{BALUT}
\alias{BALUT}
\title{Creates lookup tables for HLA Class II allele-peptide pairs based on binding affinity strength}
\usage{
BALUT(wdc2 = tempdir(), threshold = "relaxed")
}
\arguments{
\item{wdc2}{path to directory with Class II peptide binding affinity results. 
Default is set to the temp directory.}

\item{threshold}{the threshold to use for filtering NetCleave predictions.
A relaxed threshold filters peptides that are likely to be generated,
whereas a stringent threshold filters peptides that are highly likely to be generated.
Options are 'relaxed' or 'stringent'. Default is set to 'relaxed'.}
}
\value{
The full path for the Class II look up table
}
\description{
Creates a lookup table for the total number of binders, strong binders, weak binders,
and weak and strong binders for Class II CIWD alleles based on binding affinity
results from run_netmhc2pan(). Strong Binders are categorized as 
peptide-allele pairs being less than 2.0 Rank, and Weak binders as less than or 
equal to 10.0 Rank, where Rank is the rank of the predicted binding score 
compared to a set of random natural peptides. Binding strength Rank value thresholds were
obtained from Reynisson et.al "NetMHCpan-4.1 and NetMHCIIpan-4.0: improved 
predictions of MHC antigen presentation by concurrent motif deconvolution and 
integration of MS MHC eluted ligand data" 2020, W449– W454m https://doi.org/10.1093/nar/gkaa379.
The final output is output to  the user's working directory, named 
LookUpTable_ClassII_<variant>_<threshold>_<random alphanumeric sequence>.csv', where <variant> is the 
SARS-CoV2 variant and <threshold> is the threshold selected. If the bundled SARS-CoV-2 FASTA files are used, the variant
will be 'ref'.
}
