% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{beta_FASTA}
\alias{beta_FASTA}
\title{SARS-CoV-2 Beta Sequence FASTA Files}
\format{
a folder with 26 FASTA files containing FASTA sequences for the SARS-CoV-2
Beta proteome.
}
\source{
https://www.ncbi.nlm.nih.gov/nuccore/MW598419.1
}
\description{
A folder containing FASTA files the SARS-CoV-2 Beta proteome, which includes the
Envelope, Membrane, Nucleocapsid, Spike, nsp1-nsp11, nsp13-16, ORF3A, ORF6, ORF8, 
ORF10, ORF7a, ORF7b, and RdRp proteins. A custom script was used to scrape the 
FASTA sequences from https://www.ncbi.nlm.nih.gov/nuccore/MW598419.1. Note these
are user submitted sequences to NCBI, and are not officially published by NCBI
}
