% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1.R
\name{C1}
\alias{C1}
\title{Class I Pipeline Wrapper Function}
\usage{
C1(c1_predict_method, fasta_files, chop_path, mhcpan_path, dev, ds, c1_file)
}
\arguments{
\item{c1_predict_method}{Binding Affinity Prediction method used for Class I. Accepted values
are 'mhcflurry' or 'netmhcpan'. Default is set to 'netmhcpan'.}

\item{fasta_files}{FASTA file for a given protein. Default 
value is set to the inst/extdata/ref_FASTA folder, which contains FASTA files
for the SARS-CoV2 reference proteins. If the user is using other FASTA files,
the full path to the file must be provided. FASTA files should only contain
ONE protein sequence per file if Class II evaluation is desired. Please read
the vignette for more details.}

\item{chop_path}{The path to the user's installation of NetChop}

\item{mhcpan_path}{The path to the user's installation of NetMHCpan}

\item{dev}{A Boolean value for whether the function is being executed for development
and testing purposes. Default is set to FALSE. This value should NEVER be altered
by the user.}

\item{ds}{a BIGDAWG formatted HLA genotype dataset, where there is 
an identifying subject column (SID), a phenotype, where control = 0 and 
case = 1 (Disease), and two columns per HLA locus. Please see the vignette
for an example of a BIGDAWG formatted dataset.}

\item{c1_file}{Temp file for Class I pipeline log}
}
\value{
The full path for the Class I look up table. Otherwise, a fatal logging message is returned.
}
\description{
Generates peptide binding affinity scores for Class I and the reference SARS-CoV-2 strain.
Protein sequences for the reference SARS-COV-2 strain are bundled in the package. 
Users can also use COBRA for other virus sequences.
}
\note{
The user's working directory MUST be set to where the input FASTA files are located if 
the bundled SARS-CoV2 FASTA files are not being used.
}
