% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C1BAP.R
\name{C1BAP}
\alias{C1BAP}
\title{Class I Binding Affinity Predictor}
\usage{
C1BAP(filename, lookup_file)
}
\arguments{
\item{filename}{A BIGDAWG formatted dataset, which should have an identifying
subject column, a phenotype, where control = 0 and case = 1, and two columns
per Class I HLA locus. Please see the vignette for an example of a BIGDAWG
formatted dataset.
are 'mhcflurry' or 'netmhcpan'}

\item{lookup_file}{Unique name for the look up table generated from LUG()}
}
\value{
A dataframe containing peptide binding affinity predictions for Class I genotypes present in the provided BIGDAWG formatted dataset.
If the lookup table provided does not exist, '' is returned.
}
\description{
Compiles a comprehensive list of All Binders, Strong Binders, Weak Binders, and 
Weak and Strong Binders for the genotypes present in a BIGDAWG formatted dataset
for Class I HLA alleles.
}
