% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2.R
\name{C2}
\alias{C2}
\title{Class II Pipeline Wrapper Function}
\usage{
C2(
  fasta_files,
  netcleave_path,
  threshold,
  netmhcIIpanpath,
  dev,
  ds_filename,
  c2_file
)
}
\arguments{
\item{fasta_files}{FASTA file for a given protein. Default 
value is set to the inst/extdata/ref_FASTA folder, which contains FASTA files
for the SARS-CoV2 reference proteins. If the user is using other FASTA files,
the full path to the file must be provided. FASTA files should only contain
ONE protein sequence per file if Class II evaluation is desired. Please read
the vignette for more details.}

\item{netcleave_path}{The path to the cloned NetCleave repo.}

\item{threshold}{the threshold to use for filtering NetCleave predictions.
A relaxed threshold filters peptides that are likely to be generated,
whereas a stringent threshold filters peptides that are highly likely to be generated.
Options are 'relaxed' or 'stringent'. Default is set to 'relaxed'.}

\item{netmhcIIpanpath}{The path to the user's installation of NetMHCIIpan}

\item{dev}{A Boolean value for whether the function is being executed for development
and testing purposes. Default is set to FALSE. This value should NEVER be altered
by the user.}

\item{ds_filename}{a BIGDAWG formatted HLA genotype dataset, where there is 
an identifying subject column (SID), a phenotype, where control = 0 and 
case = 1 (Disease), and two columns per HLA locus. Please see the vignette
for an example of a BIGDAWG formatted dataset.}

\item{c2_file}{Temp file for Class II pipeline log}
}
\value{
The full path for the Class II look up table. Otherwise, a fatal logging message is returned.
}
\description{
Generates peptide binding affinity scores for Class II and the reference SARS-CoV-2 strain.
Protein sequences for the reference SARS-COV-2 strain are bundled in the package. 
Users can also use COBRA for other virus sequences.
}
\note{
The user's working directory MUST be set to where the input FASTA files are located if 
the bundled SARS-CoV2 FASTA files are not being used.
}
