% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2BAP.R
\name{C2BAP}
\alias{C2BAP}
\title{Class II Binding Affinity Predictor}
\usage{
C2BAP(
  filename,
  excludeDQforbidden = TRUE,
  excludeDPforbidden = TRUE,
  dpa_rule = "50",
  lookup_file
)
}
\arguments{
\item{filename}{A BIGDAWG formatted dataset, which should have an identifying
subject column, a phenotype, where control = 0 and case = 1, and two columns
per Class II HLA locus, except in the case of DRB3/4/5. Please see the vignette
for an example of a BIGDAWG formatted dataset.}

\item{excludeDQforbidden}{A Boolean value for whether forbidden DQ heterodimers should
be excluded from score evaluation. Default is set to TRUE.}

\item{excludeDPforbidden}{A Boolean value for whether forbidden DP heterodimers should
be excluded from score evaluation. Default is set to TRUE.}

\item{dpa_rule}{Dimerization rule to implement. Options include "31" or "50".}

\item{lookup_file}{Unique name for the look up table generated from BALUT()}
}
\value{
A dataframe containing peptide binding affinity predictions for Class II genotypes present in the provided BIGDAWG formatted dataset
'If the lookup table provided does not exist, '' is returned.
}
\description{
Compiles a comprehensive list of All Binders, Strong Binders, Weak Binders, and 
Weak and Strong Binders for the genotypes present in a BIGDAWG formatted dataset
for Class II HLA alleles. For DQ and DP, all possible heterodimer combinations
are generated from the BIGDAWG dataset, since DP and DQ haplotypes cannot be 
inferred. 
Forbidden DQ and DP heterodimers can be excluded during score generation. 
Forbidden DQ heterodimer rules are insert stuff here: <citation>
Forbidden DP heterodimer rules are based on peptide positions 31 and 50 (S.J. Mack, Personal Communication).
Users must select which DPA rule they would like implemented. 
Permissible heterodimers under the '31' rule include: DPA1 = Q, DPB1 = E OR if DPA1 = M, DPB1 = G.
Permissible heterodimers under the '50' rule include: DPA1 = R, DPB1 = E OR if DPA1 = Q, DPB1 = G.
}
\references{
Position 31 rules: Hollenbach, J. et.al (2012). A combined DPA1~DPB1 amino acid epitope is the primary unit of selection on the HLA-DP heterodimer. Immunogenetics, 64(8), 559–569. https://doi.org/10.1007/s00251-012-0615-3

Position 50 rules:  Mack, S.J., Personal Communication
}
