% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choppR.R
\name{choppR}
\alias{choppR}
\title{Runs NetChop on input FASTA files}
\source{
NetChop v 3.1, downloaded from https://services.healthtech.dtu.dk/services/NetMHCpan-4.1/
}
\usage{
choppR(
  fastafile = list.files(system.file("extdata/ref_FASTA", package = "COBRA"), full.names
    = TRUE),
  netchop_path
)
}
\arguments{
\item{fastafile}{FASTA file for a given protein or group of proteins. Default 
value is set to the inst/extdata/ref_FASTA folder, which contains FASTA files
for the SARS-CoV2 reference proteins}

\item{netchop_path}{The path to the user's installation of NetChop}
}
\value{
NULL. Otherwise, a string detailing the error that occurred.
}
\description{
Runs NetChop on input FASTA files to predict likelihood of cleavage at a given
amino acid position. The NetChop output, named '<file>_chop.txt', where <file>
is the name of the input FASTA file, is output to a temp directory. The file is 
further processed, where extraneous text and white space are removed, and is output
to the temp directory as '<file>_chopadj.txt'. Finally, the processed file is
read into R, where column headers for multi-protein files are removed, row 
numbers are added as a column, and output to the temp directory as '<file>_netchopfinal.txt'.
NetChop parameters: threshold for specifying weak and strong binders was set to 0.1, and the
C-terminus prediction method was used.
}
\note{
The user's working directory MUST be set to where the input FASTA files are located if 
the bundled SARS-CoV2 FASTA files are not being used.

NetChop MUST be downloaded and the path to its executable MUST be set up properly in order
for this function to work properly. In order to set up the full path to NetChop,
open the executable in a text editor, and enter the path to your NetChop download
where it says '# full path to the NetMHCpan 4.0 directory (mandatory)'.
}
\examples{
#Example to run NetChop on a FASTA file named 'EMN_ref.faa'. Path should end in 
#/netchop-3.1/netchop'
\donttest{choppR("EMN_ref.faa", "path/to/NetChop/executable")}

}
\references{
Nielsen M, Lundegaard C, Lund O, and Kesmir C. The role of the proteasome in generating cytotoxic T cell epitopes: Insights obtained from improved predictions of proteasomal cleavage. Immunogenetics., 57(1-2):33-41, 2005.
}
