% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{CIWD_C1_mhcflurry}
\alias{CIWD_C1_mhcflurry}
\title{HLA Class I CIWD Allele files}
\format{
a folder with 46 .txt files containing HLA Class I CIWD alelles
}
\source{
CIWD_CLASSI_Palleles-all_loci-2020320-ihws-website.xlsx from https://www.ihiw18.org/component-immunogenetics/download-common-and-well-documented-alleles-3-0/
}
\description{
A folder containing all Common, Intermediate, Well-Documented (CIWD) v 3.0.0 Class I
HLA Alleles. These files will be used as input if mhcflurry is selected as the
Binding Affinity Predictor for Class I alleles. The source file was parsed
to only include CIWD alleles, to only alleles allowed by mhcflurry,
and non-expression alleles (N, S, C, A and Q) were filtered out. The comprehensive
list of Class I CIWD HLA alleles was broken up to generate files of 45 alleles per file.
These files are formatted differently than CIWD_C1_netmhcpan files in that they have spaces 
instead of commas as the delimiter, and do not contain HLA-C12:139, as this allele is not supported by NetCleave.
}
\references{
Hurley et. al. "Common, intermediate and well-documented HLA alleles in world populations: CIWD version 3.0.0". HLA (2020), Volume 95, Issue 6, pgs 503-637
}
