% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\docType{data}
\name{CIWD_C2}
\alias{CIWD_C2}
\title{HLA Class II CWD/CIWD alleles}
\format{
a folder with 105 .txt files containing HLA Class II CIWD/CWD alleles
}
\source{
CIWD_CLASS_Palleles-all_loci-2020320-ihws-website.xlsx from https://www.ihiw18.org/component-immunogenetics/download-common-and-well-documented-alleles-3-0/

cwd200.xls from https://www.uvm.edu/~igdawg/cwd.html
}
\description{
A folder containing all CIWD v 3.0.0 DRB1, DQB1, and DPB1 alleles, and CWD v 2.0.0
CWD v 2.0.0 DPA1 and DQA1 alleles. The source files were parsed
to only include Class II CIWD/CWD alleles, to only alleles allowed by netMHCIIpan, 
and non-expression alleles (N, S, C, A and Q) were filtered out. The 
comprehensive list of CIWD/CWD Class II HLA alleles was broken up to generate 
20 alleles per file to accommodate netMHCIIPan’s maximum allele evaluation capacity.  
The source files were also manipulated to create all possible heterodimers for DP and DQ, 
including forbidden heterodimers.
}
\references{
Hurley et. al. "Common, intermediate and well-documented HLA alleles in world populations: CIWD version 3.0.0". HLA (2020), Volume 95, Issue 6, pages 503-637

Mack et. al. "Common and well-documented HLA alleles: 2012 update to the CWD catalogue". Tissue Antigens (2013), Volume 8, Issue 4 pages 194-203
}
