% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaveR.R
\name{cleaveR}
\alias{cleaveR}
\title{Runs NetCleave on input FASTA files}
\usage{
cleaveR(
  fastafiles = list.files(system.file("extdata/ref_FASTA", package = "COBRA"), full.names
    = TRUE),
  netcleave_path
)
}
\arguments{
\item{fastafiles}{FASTA file for a given protein. Default 
value is set to the inst/extdata/ref_FASTA folder, which contains FASTA files
for the SARS-CoV2 reference proteins. If the user is using other FASTA files,
the full path to the file must be provided. FASTA files should only contain
ONE protein sequence per file.}

\item{netcleave_path}{The path to the cloned NetCleave repo.}
}
\value{
TRUE. Otherwise, a string detailing the error that occurred.
}
\description{
Runs NetCleave on input FASTA files to predict likelihood of
lysosomal cleavage at the fourth amino acid position. The NetCleave output,
named '<file>_NetCleave.csv', where <file> is the name of the input FASTA file, is
output to the temp directory.
}
\note{
NetCleave is not available via pip installation. Python scripts for NetCleave
were forked from https://github.com/BSC-CNS-EAPM/. Some scripts (NetCleave.py and
and cleavage_site_generator.py) were modified to accept an output path (--td) as 
an argument. The temp directory path is provided for the NetCleave predictions 
to be output to. Users MUST clone the Github repo with the modified 
scripts. The repo can be cloned by executing 'git clone https://github.com/liviatran/NetCleave.git'.
Link to the forked repo with modified scripts: https://github.com/liviatran/NetCleave
}
\references{
Amengual-Rigo, P., Guallar, V. NetCleave: an open-source algorithm for predicting C-terminal antigen processing for MHC-I and MHC-II. Sci Rep 11, 13126 (2021). https://doi.org/10.1038/s41598-021-92632-y
}
