% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{COBRA-package}
\alias{COBRA}
\alias{COBRA-package}
\title{COBRA: COrrection of BAtch Effect for Single-Cell RNA-Seq Data}
\description{
Removes unwanted batch variation from single-cell RNA-seq expression
matrices while preserving biological signal. When cell-type labels are
unavailable, pseudo-cell types are estimated automatically via
iterative clustering with elbow-based PC selection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wonlab-healthstat/COBRA}
  \item Report bugs at \url{https://github.com/wonlab-healthstat/COBRA/issues}
}

}
\author{
\strong{Maintainer}: Sujin Seo \email{sujin91kr@gmail.com}

}
