% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configure_python.R
\name{configure_python}
\alias{configure_python}
\title{Configure system with required modules to run Python code, either via a virtual environment or the user's machine}
\usage{
configure_python()
}
\value{
""Previously configured virtual environment activated successfully!" or a string detailing the error that occurred.
}
\description{
A properly configured Python environment is required to run COBRA for Class II HLA alleles, 
or if the user wishes to use mhcflurry as the peptide binding affinity predictor. 
A virtual environment named "COBRA" will be created, and modules required to run
mhcflurry, or the Class II pipeline, will be installed to the virtual environment. Once successful, the project's
Python interpreter will be set to the virtual environment's Python interpreter.
Troubleshooting advice for virtual environment activation and installation can be found in the vignette.
}
\note{
Python 3 must be installed on the user's system.

For internal use only.
}
