% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerizeI.R
\name{kmerizeI}
\alias{kmerizeI}
\title{Generate possible generated peptides during peptide processing for HLA Class I alleles}
\usage{
kmerizeI(
  FASTAfile = list.files(system.file("extdata/ref_FASTA", package = "COBRA"), full.names
    = TRUE),
  ba_predict
)
}
\arguments{
\item{FASTAfile}{FASTA file for a given protein or group of proteins. Default 
value is set to the inst/extdata/ref_FASTA folder, which contains FASTA files
for the SARS-CoV2 reference proteins}

\item{ba_predict}{Binding Affinity Prediction method used for Class I. Accepted values
are 'mhcflurry' or 'netmhcpan'}
}
\value{
The names of the FASTA files with generated peptides. Otherwise, "No peptides" is returned.
}
\description{
Generates a list of estimated 8-12 residue long peptides based on potential
split site information obtained from NetChop for Class I HLA alleles. Possible
ending amino acids are indicated by potential split sites, and possible beginning 
amino acid positions are those directly after a split position. The list is 
output to the temp directory. If 'netMHCpan' is specified as the binding affinity
predictor, the output will be named 'ClassI_Peptides_<file>_.txt', 
where <file> is the name of the FASTA file. If 'mhcflurry' is selected, 
the output will be named '<file>_pep.txt'.
}
\note{
The user's working directory MUST be set to where the input FASTA files are located if 
the bundled SARS-CoV2 FASTA files are not being used.
}
