% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmerizeII.R
\name{kmerizeII}
\alias{kmerizeII}
\title{Filters NetCleave results to likely to be generated epitopes for Class II HLA alleles}
\usage{
kmerizeII(
  fasta = list.files(system.file("extdata/ref_FASTA", package = "COBRA"), full.names =
    TRUE),
  threshold = "relaxed"
)
}
\arguments{
\item{fasta}{FASTA file for a given protein. Default 
value is set to the inst/extdata/ref_FASTA folder, which contains FASTA files
for the SARS-CoV2 reference proteins. If the user is using other FASTA files,
the full path to the file must be provided. FASTA files should only contain
ONE protein sequence per file.}

\item{threshold}{the threshold to use for filtering NetCleave predictions.
A relaxed threshold filters peptides that are likely to be generated,
whereas a stringent threshold filters peptides that are highly likely to be generated.
Options are 'relaxed' or 'stringent'. Default is set to 'relaxed'.}
}
\value{
The names of the FASTA files with generated peptides. Otherwise, "No peptides" is returned.
}
\description{
Given a FASTA file, NetCleave generates all possible epitopes (13-17mers) for 
Class II HLA alleles and scores the likelihood of generation. Peptides with 
scores greater than the specified threshold of epitope generation are
filtered to be included in the output file, where the 'relaxed' threshold is > 0.5, and the
'stringent' threshold is > 0.6. The filtered file is output
to the temp directory, named, named 'ClassII_Peptides_<file>_.txt', 
where <file> is the name of the FASTA file.
}
\note{
The user's working directory MUST be set to where the input FASTA files are located if 
the bundled SARS-CoV2 FASTA files are not being used.
}
