% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobra_functions.R
\name{lm.fit.large.new}
\alias{lm.fit.large.new}
\title{Large matrix fit (method 2, with Rfast transpose)}
\usage{
lm.fit.large.new(xpp, dat)
}
\description{
Large matrix fit (method 2, with Rfast transpose)
}
\keyword{internal}
