% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LUG.R
\name{LUG}
\alias{LUG}
\title{Creates lookup tables for HLA Class I allele-peptide pairs based on binding affinity strength}
\usage{
LUG(wdc1 = tempdir(), ba_predict)
}
\arguments{
\item{wdc1}{path to directory with Class I peptide binding affinity results. 
Default is set to the temp directory.}

\item{ba_predict}{Binding Affinity Prediction method used for Class I. Accepted values
are 'mhcflurry' or 'netmhcpan'}
}
\value{
The full path for the Class I look up table
}
\description{
Creates a lookup table for the total number of binders, strong binders, weak binders,
and weak and strong binders for Class I CIWD alleles based on binding affinity
results from run_netmhcpan() or run_mhcflurry(). Strong Binders are categorized as 
peptide-allele pairs being less than 0.5 Rank, and Weak binders as less than or 
equal to 2.0 Rank, where Rank is the rank of the predicted binding score 
compared to a set of random natural peptides. Binding strength Rank value thresholds were
obtained from Reynisson et.al "NetMHCpan-4.1 and NetMHCIIpan-4.0: improved 
predictions of MHC antigen presentation by concurrent motif deconvolution and 
integration of MS MHC eluted ligand data" 2020, W449– W454m https://doi.org/10.1093/nar/gkaa379.
Only peptide-allele pairs with
at least loose binding peptide-allele pairs (less than 500 nm) will be evaluated. 
The final output is output to  the user's working directory, named 
LookUpTable_ClassI_<variant>_<predict-method>_<random alphanumeric sequence>.csv', where <variant> is the 
SARS-CoV2 variant, and <predict_method> is the binding affinity prediction 
method used. If the bundled SARS-CoV-2 FASTA files are used, the variant will be 'ref'.
}
