% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mhcflurry.R
\name{run_mhcflurry}
\alias{run_mhcflurry}
\title{Runs mhcflurry on allele-peptide data}
\source{
mhcflurry Github Repo: https://github.com/openvax/mhcflurry

T. J. O’Donnell, et al. “MHCflurry 2.0: Improved pan-allele prediction of MHC I-presented peptides by incorporating antigen processing,” Cell Systems, 2020. https://doi.org/10.1016/j.cels.2020.06.010
}
\usage{
run_mhcflurry(
  FASTAfile = list.files(system.file("extdata/ref_FASTA", package = "COBRA"), full.names
    = TRUE),
  dev = FALSE
)
}
\arguments{
\item{FASTAfile}{FASTA file for a given protein or group of proteins. Default 
value is set to the inst/extdata/ref_FASTA folder, which contains FASTA files
for the SARS-CoV2 reference proteins}

\item{dev}{A Boolean value for whether the function is being executed for development
and testing purposes. Default is set to FALSE. This value should NEVER be altered
by the user.}
}
\value{
TRUE. Otherwise, a string detailing the error that occurred.
}
\description{
Runs mhcflurry to obtain peptide binding affinity for CIWD HLA Class I alleles and possible generated peptides.
CIWD files formatted specifically for mhcflurry are bundled in the package. mhcflurry
will run in the COBRA venv if venv was set to TRUE in configure_python_mhcflurry(). 
mhcflurry results are output to the temporary directory, and are named 'ClassICIWD_<n>_<file>_pep_flur.csv',
where <n> is number for the corresponding CIWD file and <file> is the name of the FASTA file.
The jobs are run in parallel, with the optimal number of jobs determined based on the user's
machine's processing power.
}
\note{
The user's working directory MUST be set to where the input FASTA files are located if 
the bundled SARS-CoV2 FASTA files are not being used.

GNU Parallel MUST be installed in order for parallel computation to 
occur. It can be installed from https://www.gnu.org/software/parallel/, or via
homebrew with 'brew install parallel'.
}
