% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_netmhc2pan.R
\name{run_netmhc2pan}
\alias{run_netmhc2pan}
\title{Run NetMHCpan on Class II HLA allele-peptide data}
\source{
NetMHCIIpan 4.1, downloaded from https://services.healthtech.dtu.dk/services/NetMHCIIpan-4.0/

Reynisson B, Barra C, Kaabinejadian S, Hildebrand WH, Peters B, Nielsen M. Improved Prediction of MHC II Antigen Presentation through Integration and Motif Deconvolution of Mass Spectrometry MHC Eluted Ligand Data. J Proteome Res. 2020 Jun 5;19(6):2304-2315. doi: 10.1021/acs.jproteome.9b00874. Epub 2020 Apr 30. PMID: 32308001.
}
\usage{
run_netmhc2pan(
  faaname = list.files(system.file("extdata/ref_FASTA", package = "COBRA"), full.names =
    TRUE),
  netMHC2pan_path,
  dev = FALSE
)
}
\arguments{
\item{faaname}{FASTA file for a given protein or group of proteins. Default 
value is set to the inst/extdata/ref_FASTA folder, which contains FASTA files
for the SARS-CoV2 reference proteins}

\item{netMHC2pan_path}{The path to the user's installation of NetMHCIIpan}

\item{dev}{A Boolean value for whether the function is being executed for development
and testing purposes. Default is set to FALSE. This value should NEVER be altered
by the user.}
}
\value{
NULL. Otherwise, a string detailing the error that occurred.
}
\description{
Runs NetMHCIIpan to obtain peptide binding affinity for CIWD/CWD HLA Class II alleles and possible generated peptides.
CIWD files formatted specifically for NetMHCIIpan are bundled in the package. NetMHCpan results are output to the temporary
directory, and are named 'ClassI_Peptides_<file>_ClassICIWD_<n>_BindAff.txt', 
where <file> is the name of the FASTA file, and <n> is number for the corresponding 
CIWD file. 53 jobs are executed in parallel for faster computation.
}
\note{
The user's working directory MUST be set to where the input FASTA files are located if 
the bundled SARS-CoV2 FASTA files are not being used.

NetMHCIIpan MUST be downloaded and the path to its executable MUST be set up properly in order
for this function to work properly. NetMHCIIpan v 4.1 was used at the time of 
COBRA development. In order to set up the full path to NetMHCIIpan,
open the executable in a text editor, and enter the path to your NetMHCIIpan download
where it says '# full path to the NetMHCIIpan 4.1 directory (mandatory)'.

GNU Parallel MUST be installed in order for parallel computation to 
occur. It can be installed from https://www.gnu.org/software/parallel/, or via
homebrew with 'brew install parallel'.
}
