% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_netmhcpan.R
\name{run_netmhcpan}
\alias{run_netmhcpan}
\title{Run NetMHCpan on Class I HLA allele-peptide data}
\source{
NetMHCpan 4.1, downloaded from https://services.healthtech.dtu.dk/services/NetMHCpan-4.1/

Reynisson B, Alvarez B, Paul S, Peters B, Nielsen M. NetMHCpan-4.1 and NetMHCIIpan-4.0: improved predictions of MHC antigen presentation by concurrent motif deconvolution and integration of MS MHC eluted ligand data, Nucleic Acids Research, Volume 48, Issue W1, 02 July 2020, Pages W449– W454, https://doi.org/10.1093/nar/gkaa379
}
\usage{
run_netmhcpan(
  fastafilename = list.files(system.file("extdata/ref_FASTA", package = "COBRA"),
    full.names = TRUE),
  netmhcpan_path,
  dev = FALSE
)
}
\arguments{
\item{fastafilename}{FASTA file for a given protein or group of proteins. Default 
value is set to the inst/extdata/ref_FASTA folder, which contains FASTA files
for the SARS-CoV2 reference proteins}

\item{netmhcpan_path}{The path to the user's installation of NetMHCpan}

\item{dev}{A Boolean value for whether the function is being executed for development
and testing purposes. Default is set to FALSE. This value should NEVER be altered
by the user.}
}
\value{
NULL. Otherwise, a string detailing the error that occurred.
}
\description{
Runs NetMHCpan to obtain peptide binding affinity for CIWD HLA Class I alleles and possible generated peptides.
CIWD files formatted specifically for NetMHCpan are bundled in the package. NetMHCpan results are output to the temporary
directory, and are named 'ClassI_Peptides_<file>_ClassICIWD_<n>_BindAff.txt', 
where <file> is the name of the FASTA file, and <n> is number for the corresponding 
CIWD file. 46 jobs are executed in parallel for faster computation.
}
\note{
The user's working directory MUST be set to where the input FASTA files are located if 
the bundled SARS-CoV2 FASTA files are not being used.

NetMHCpan MUST be downloaded and the path to its executable MUST be set up properly in order
for this function to work properly. NetMHCpan 4.0 was used at the time of COBRA
development. In order to set up the full path to NetMHCpan, open the executable
in a text editor, and enter the path to your NetMHCpan download where it says 
# full path to the NetMHCpan 4.0 directory (mandatory)'.

GNU Parallel MUST be installed in order for parallel computation to 
occur. It can be installed from https://www.gnu.org/software/parallel/, or via
homebrew with 'brew install parallel'.
}
