% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobra_functions.R
\docType{package}
\name{COBRA-package}
\alias{COBRA}
\alias{COBRA-package}
\title{COBRA: COBRA: A Batch Effect Correction Method for scRNA}
\description{
Provides a batch effect correction method for single-cell RNA-seq data.
}
\author{
\strong{Maintainer}: Sujin FamilyName \email{your_email@domain.com}

}
\keyword{internal}
