% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cobra_functions.R
\name{EstimatePseudoCell}
\alias{EstimatePseudoCell}
\title{Estimate Pseudo Cells}
\usage{
EstimatePseudoCell(
  data_mat,
  batch,
  ncl_min = 2,
  ncl_max = 10,
  seed = 1,
  r_random_cell = 0.01,
  n_random_cell = 1000,
  pheno = NULL
)
}
\arguments{
\item{data_mat}{Expression matrix (cells x features)}

\item{batch}{A factor or vector indicating batch labels}

\item{ncl_min}{Minimum number of clusters to test}

\item{ncl_max}{Maximum number of clusters to test}

\item{seed}{Random seed}

\item{r_random_cell}{Proportion of cells for random sampling}

\item{n_random_cell}{Absolute minimum number of cells for sampling}

\item{pheno}{Additional covariates (if any)}
}
\value{
A factor of cluster memberships (pseudo-cells)
}
\description{
When \code{cell.var} is NULL, we cluster the data to estimate pseudo-cells
and iteratively remove batch effects.
}
